[VAS-894] Updated & Cleaned StationsTableColumns
