+++ src/pages/stations/list/__tests__/StationsTableColumns.test.tsx
@@ -1,6 +1,6 @@
-import {GridColumnHeaderParams, GridRenderCellParams, GridStateColDef} from '@mui/x-data-grid';
-import {cleanup} from '@testing-library/react';
-import {buildColumnDefs, renderCell, showCustomHeader, showStationID, showStatus,} from '../StationsTableColumns';
+import { GridColDef, GridRenderCellParams, GridStateColDef } from '@mui/x-data-grid';
+import { cleanup } from '@testing-library/react';
+import { showStatus, buildColumnDefs } from '../StationsTableColumns';
 
 beforeEach(() => {
   jest.spyOn(console, 'error').mockImplementation(() => {});
@@ -23,157 +23,126 @@ describe('<StationsTableColumns />', () => {
 
     const colDefMocked: GridStateColDef<any, any, any> = {
       computedWidth: 0,
-      field: 'station_code',
-      type: 'string',
+      field: 'stationCode',
+      type: '',
       hasBeenResized: undefined,
       groupPath: undefined,
-      headerName: 'Station Name',
-    };
-
-    const colDefMockedStatus: GridStateColDef<any, any, any> = {
-      computedWidth: 0,
-      field: 'endabled',
-      type: 'string',
-      hasBeenResized: undefined,
-      groupPath: undefined,
-      headerName: 'status',
-    };
-
-    const colDefMockedBroker: GridStateColDef<any, any, any> = {
-      computedWidth: 0,
-      field: 'broker_description',
-      type: 'string',
-      hasBeenResized: undefined,
-      groupPath: undefined,
-      headerName: 'Creation Date',
-    };
-
-    const customHeaderStation: GridColumnHeaderParams = {
-      field: 'station_code',
-      colDef: colDefMocked,
-    };
-
-    const customHeaderBroker: GridColumnHeaderParams = {
-      field: 'broker_description',
-      colDef: colDefMockedBroker,
-    };
-
-    const customHeaderStatus: GridColumnHeaderParams = {
-      field: 'enabled',
-      colDef: colDefMockedStatus,
+      headerName: 'stationCode',
     };
 
     const params: GridRenderCellParams<any, any, any> = {
       value: 'some value',
       row: {
-        station_code: '123456',
-        status: 'ACTVE',
+        stationCode: '123456',
       },
       api: undefined,
-      id: '1',
-      field: 'station_code',
+      id: '',
+      field: '',
       rowNode: rowNode[0],
-      colDef: colDefMocked,
+      // @ts-ignore
+      colDef: colDefMocked[0],
       cellMode: 'edit',
       hasFocus: false,
       tabIndex: 0,
       getValue: () => jest.fn(),
     };
 
-    const paramsBroker: any = {
-      value: 'broker_description',
-      row: { broker_description: 'broker_description' },
-      field: 'broker_description',
-      api: null,
-      getValue: () => '',
-      colDef: colDefMockedBroker,
-      id: '',
-      rowNode: rowNode[0],
-      cellMode: 'view',
-      hasFocus: true,
-      tabIndex: 0,
-    };
-
-    const paramsStatus: any = {
-      value: 'status',
-      row: { enabled: 'ACTIVE' },
-      field: 'enabled',
-      api: null,
-      getValue: () => '',
-      colDef: colDefMockedStatus,
-      id: '',
-      rowNode: rowNode[0],
-      cellMode: 'view',
-      hasFocus: true,
-      tabIndex: 0,
-    };
-
-    // const ArrayBuildColumnDefs = [
-    //   {
-    //     field: 'station_code',
-    //     cellClassName: 'justifyContentBold',
-    //     headerName: 'Station Name',
-    //     align: 'left',
-    //     headerAlign: 'left',
-    //     width: 403,
-    //     editable: false,
-    //     disableColumnMenu: true,
-    //     renderHeader: showCustomHeader,
-    //     renderCell: (params: any) => showStationCode(params),
-    //     sortable: true,
-    //     flex: 4,
-    //   },
-    //   {
-    //     field: 'broker_description',
-    //     cellClassName: 'justifyContentNormal',
-    //     headerName: 'Creation Date',
-    //     align: 'left',
-    //     headerAlign: 'left',
-    //     width: 404,
-    //     editable: false,
-    //     disableColumnMenu: true,
-    //     renderHeader: showCustomHeader,
-    //     renderCell: (params) => renderCell(params, undefined),
-    //     sortable: false,
-    //     flex: 4,
-    //   },
-    //   {
-    //     field: 'enabled',
-    //     cellClassName: 'justifyContentNormal',
-    //     headerName: 'Status',
-    //     align: 'left',
-    //     headerAlign: 'left',
-    //     width: 404,
-    //     editable: false,
-    //     disableColumnMenu: true,
-    //     renderHeader: showCustomHeader,
-    //     renderCell: (params) => renderCell(params, undefined),
-    //     sortable: false,
-    //     flex: 4,
-    //     enabled: true,
-    //   },
-    //   {
-    //     field: 'actions',
-    //     cellClassName: 'justifyContentNormalRight',
-    //     type: 'actions',
-    //     headerName: '',
-    //     align: 'center',
-    //     hideSortIcons: true,
-    //     disableColumnMenu: true,
-    //     editable: false,
-
-    //     getActions: () => <React.Fragment></React.Fragment>,
-    //     sortable: false,
-    //     flex: 1,
-    //   },
-    // ] as Array<GridColDef>;
+    const ArrayBuildColumnDefs = [
+      {
+        field: 'stationCode',
+        cellClassName: 'justifyContentBold',
+        headerName: 'Station Name',
+        align: 'left',
+        headerAlign: 'left',
+        minWidth: 485,
+        editable: false,
+        disableColumnMenu: true,
+        renderHeader: expect.any(Function),
+        renderCell: expect.any(Function),
+        sortable: true,
+        flex: 4,
+      },
+      {
+        field: 'createdAt',
+        cellClassName: 'justifyContentNormal',
+        headerName: 'Creation Date',
+        align: 'left',
+        headerAlign: 'left',
+        maxWidth: 200,
+        editable: false,
+        disableColumnMenu: true,
+        renderHeader: expect.any(Function),
+        renderCell: expect.any(Function),
+        sortable: false,
+        flex: 4,
+      },
+      {
+        field: 'modifiedAt',
+        cellClassName: 'justifyContentNormal',
+        headerName: 'Last Edit Date',
+        align: 'left',
+        headerAlign: 'left',
+        maxWidth: 200,
+        editable: false,
+        disableColumnMenu: true,
+        renderHeader: expect.any(Function),
+        renderCell: expect.any(Function),
+        sortable: false,
+        flex: 4,
+      },
+      {
+        field: 'activationDate',
+        cellClassName: 'justifyContentNormal',
+        headerName: 'Activation Date',
+        align: 'left',
+        headerAlign: 'left',
+        maxWidth: 220,
+        editable: false,
+        disableColumnMenu: true,
+        renderHeader: expect.any(Function),
+        renderCell: expect.any(Function),
+        sortable: false,
+        flex: 4,
+      },
+      {
+        field: 'wrapperStatus',
+        cellClassName: 'justifyContentNormal',
+        headerName: 'Status',
+        align: 'left',
+        headerAlign: 'left',
+        width: 145,
+        editable: false,
+        disableColumnMenu: true,
+        renderHeader: expect.any(Function),
+        renderCell: expect.any(Function),
+        sortable: false,
+        flex: 4,
+      },
+      {
+        field: 'actions',
+        cellClassName: 'justifyContentNormalRight',
+        headerName: '',
+        align: 'center',
+        disableColumnMenu: true,
+        editable: false,
+        flex: 1,
+        getActions: expect.any(Function),
+        hideSortIcons: true,
+        sortable: false,
+        type: 'actions',
+      },
+    ] as Array<GridColDef>;
 
     const mockTFunction = (key: string) => {
       switch (key) {
         case 'stationsPage.stationsTableColumns.headerFields.name':
           return 'Station Name';
-        case 'stationsPage.stationsTableColumns.headerFields.description':
+        case 'stationsPage.stationsTableColumns.headerFields.creationDate':
           return 'Creation Date';
+        case 'stationsPage.stationsTableColumns.headerFields.lastEditDate':
+          return 'Last Edit Date';
+        case 'stationsPage.stationsTableColumns.headerFields.activationDate':
+          return 'Activation Date';
         case 'stationsPage.stationsTableColumns.headerFields.status':
           return 'Status';
         default:
@@ -181,45 +150,7 @@ describe('<StationsTableColumns />', () => {
       }
     };
 
-    buildColumnDefs(mockTFunction);
-
-    showStationID(params);
     showStatus(params);
-    showCustomHeader(customHeaderStation);
-    renderCell(params);
-
-    showStationID(paramsBroker);
-    showStatus(paramsBroker);
-    showCustomHeader(customHeaderBroker);
-    renderCell(paramsBroker, undefined);
-
-    showStationID(paramsStatus);
-    showStatus(paramsStatus);
-    showCustomHeader(customHeaderStatus);
-    renderCell(paramsStatus, undefined);
-
-    // const columnDefs = buildColumnDefs(mockTFunction, onRowClickMocked);
-
-    // const renderCellMocked =
-    //   typeof renderCell !== 'undefined' ? columnDefs[0].renderCell : colDefMocked[0].field;
-
-    // const result = renderCellMocked(ArrayBuildColumnDefs[0].field);
-    // expect(result).toEqual(ArrayBuildColumnDefs[0]);
-
-    // const columnDefsBrk = buildColumnDefs(mockTFunction, onRowClickMocked);
-    // const renderCellBrkMocked =
-    //   typeof renderCell !== 'undefined' ? columnDefsBrk[1].renderCell : colDefMockedBroker[1].field;
-
-    // const resultBrk = renderCellBrkMocked(ArrayBuildColumnDefs[1].field);
-    // expect(resultBrk).toEqual(ArrayBuildColumnDefs[1]);
-
-    // const columnDefsStatus = buildColumnDefs(mockTFunction, onRowClickMocked);
-    // const renderCellStatus =
-    //   typeof renderCell !== 'undefined'
-    //     ? columnDefsStatus[2].renderCell
-    //     : colDefMockedBroker[1].field;
-
-    // const resultStatus = renderCellStatus(ArrayBuildColumnDefs[2].field);
-    // expect(resultStatus).toEqual(ArrayBuildColumnDefs[2]);
+    expect(buildColumnDefs(mockTFunction)).toEqual(ArrayBuildColumnDefs);
   });
 });
